pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "langchain-core>=0.3.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-google-genai>=4.2.0",
    "pypdf>=6.6.2",
    "reportlab>=4.4.9",
    "streamlit>=1.53.1",
    "uvicorn>=0.40.0",
    "pyyaml>=6.0.3",
]
